38. Eloquent ORM Edit,Update Data With Image & Without Image Part 2

+++ app/Http/Controllers/BrandController.php
@@ -67,42 +67,44 @@ class BrandController extends Controller
         $brand_image = $request->file('brand_image');
 
 
-        $name_gen = hexdec(uniqid());
-        $img_ext = strtolower($brand_image->getClientOriginalExtension());
-        $img_name = $name_gen.'.'.$img_ext;
-        $up_location = 'image/brand/';
-        $last_img = $up_location.$img_name;
-        $brand_image->move($up_location,$img_name);
+        if($brand_image){
+            $name_gen = hexdec(uniqid());
+            $img_ext = strtolower($brand_image->getClientOriginalExtension());
+            $img_name = $name_gen.'.'.$img_ext;
+            $up_location = 'image/brand/';
+            $last_img = $up_location.$img_name;
+            $brand_image->move($up_location,$img_name);
 
-        unlink($old_image);
-        Brand::find($id)->update([
-            'brand_name'  => $request->brand_name,
-            'brand_image' => $last_img,
-            'created_at' => Carbon::now()]);
-            return Redirect()->back()->with('success', 'Brand Updated Successfully!');
+            unlink($old_image);
+            Brand::find($id)->update([
+                'brand_name'  => $request->brand_name,
+                'brand_image' => $last_img,
+                'created_at' => Carbon::now()
+        ]);
+        return Redirect()->back()->with('success', 'Brand Image Updated Successfully!');
+        }else{
+            Brand::find($id)->update([
+                'brand_name'  => $request->brand_name,
+                'created_at' => Carbon::now()
+        ]);
+        return Redirect()->back()->with('success', 'Brand Name Updated Successfully!');
+        }
 
-        // if($brand_image){
-        //     $name_gen = hexdec(uniqid());
-        //     $img_ext = strtolower($brand_image->getClientOriginalExtension());
-        //     $img_name = $name_gen.'.'.$img_ext;
-        //     $up_location = 'image/brand/';
-        //     $last_img = $up_location.$img_name;
-        //     $brand_image->move($up_location,$img_name);
+        // $name_gen = hexdec(uniqid());
+        // $img_ext = strtolower($brand_image->getClientOriginalExtension());
+        // $img_name = $name_gen.'.'.$img_ext;
+        // $up_location = 'image/brand/';
+        // $last_img = $up_location.$img_name;
+        // $brand_image->move($up_location,$img_name);
 
-        //     unlink($old_image);
-        //     Brand::find($id)->update([
-        //         'brand_name'  => $request->brand_name,
-        //         'brand_image' => $last_img,
-        //         'created_at' => Carbon::now()
-        // ]);
-        // return Redirect()->back()->with('success', 'Brand Updated Successfully!');
-        // }else{
-        //     Brand::find($id)->update([
-        //         'brand_name'  => $request->brand_name,
-        //         'created_at' => Carbon::now()
-        // ]);
-        // return Redirect()->back()->with('success', 'Brand Updated Successfully!');
-        // }
+        // unlink($old_image);
+        // Brand::find($id)->update([
+        //     'brand_name'  => $request->brand_name,
+        //     'brand_image' => $last_img,
+        //     'created_at' => Carbon::now()]);
+        //     return Redirect()->back()->with('success', 'Brand Updated Successfully!');
+
+
 
 
     }
